refactor: extract composer view from AgentWorkspace
