+++ src/components/agent/composer/AgentComposer-types.ts
@@ -1,0 +1,147 @@
+import type { Editor as TiptapEditor } from "@tiptap/react";
+import { type ClipboardEvent, type DragEvent, type FormEvent } from "react";
+import { type AgentTaskDto, type HermesSkillInfo, type VeniceModelDto } from "../../../lib/tauri";
+import { type TextFundingModelContext } from "../../../lib/account-gate";
+import type { ScrollFade } from "../../../lib/use-scroll-fade";
+import { type ThinkingLevel } from "../../../lib/thinking-level";
+import type { TextFundingNoticeContext } from "../../account/FundingNotice";
+import { type ModelPickerFlyout } from "../../settings/ModelPickerPopover";
+import { type ComposerEditorHandle } from "./ComposerEditor";
+import { type ReportCategory } from "./reportCategory";
+import { type ReportDialogAttachment } from "../ReportDialog";
+import type { AgentAttachment } from "../agent-workspace-models";
+import { type AgentPanel } from "../agent-workspace-config";
+import type { PendingIssueReport } from "../agent-session-continuity";
+import { type QueuedAttachmentFollowUp } from "./follow-up-queue";
+import { type ComposerInputSizeWarning } from "./composer-input-helpers";
+import type * as React from "react";
+
+export type RenderAgentComposerDependencies = {
+  SESSION_BUSY_NOTICE: "June is still working on the previous message.";
+  activeGenerationCostQuality: number | undefined;
+  activePanel: AgentPanel;
+  agentScrollRef: React.MutableRefObject<HTMLDivElement | null>;
+  attachMenuOpen: boolean;
+  attachMenuRef: React.MutableRefObject<HTMLDivElement | null>;
+  attachTriggerRef: React.MutableRefObject<HTMLButtonElement | null>;
+  attachments: AgentAttachment[];
+  attachmentsRef: React.MutableRefObject<AgentAttachment[]>;
+  categoryRef: React.MutableRefObject<ReportCategory | null>;
+  composerBoxRef: React.MutableRefObject<HTMLDivElement | null>;
+  composerDraftKeyRef: React.MutableRefObject<string | null>;
+  composerEditorRef: React.MutableRefObject<ComposerEditorHandle | null>;
+  composerInSteerState: boolean;
+  composerModelFlyout: ModelPickerFlyout;
+  composerModelOpen: boolean;
+  composerModelPopoverRef: React.RefObject<HTMLDivElement>;
+  composerModelRootSearchRef: React.RefObject<HTMLInputElement>;
+  composerModelSearchRef: React.RefObject<HTMLInputElement>;
+  composerModelTriggerRef: React.RefObject<HTMLButtonElement>;
+  composerRef: React.MutableRefObject<HTMLFormElement | null>;
+  composerThinkingLevel: ThinkingLevel;
+  composerTiptapEditorRef: React.MutableRefObject<TiptapEditor | null>;
+  confirmUnrestricted: boolean;
+  creditActionsDisabledReason: string | undefined;
+  draft: string;
+  draftRef: React.MutableRefObject<string>;
+  dropActive: boolean;
+  editOversizeComposerInput: () => void;
+  fullModeDraft: boolean;
+  fullModeDraftRef: React.MutableRefObject<boolean>;
+  galleryErrors: boolean;
+  generatingImage: boolean;
+  generatingVideo: boolean;
+  generationModel: VeniceModelDto | undefined;
+  generationModelOptions: VeniceModelDto[];
+  handleComposerDragOver: (event: DragEvent<HTMLFormElement>) => void;
+  handleComposerDrop: (event: DragEvent<HTMLFormElement>) => void;
+  handleComposerPaste: (event: ClipboardEvent<HTMLFormElement>) => void;
+  handleCostQualityChange: (value: number) => void;
+  handleReportDialogSent: () => void;
+  handleSelectGenerationModel: (
+    modelId: string,
+    costQuality?: number,
+    options?: { keepOpen?: boolean; targetStoredSessionId?: string | null },
+  ) => Promise<boolean>;
+  handleSelectThinkingLevel: (level: ThinkingLevel) => Promise<void>;
+  heroMode: boolean;
+  imageModelWarningText: string;
+  imageSlashBlockedByModel: boolean;
+  importReportDialogDroppedFiles: (files: File[]) => Promise<boolean>;
+  importingFiles: boolean;
+  loadSkillCommands: (options?: { silent?: boolean }) => Promise<HermesSkillInfo[]>;
+  modelRootSearch: string;
+  modelSearch: string;
+  openComposerModelPicker: (fromSlash?: boolean) => void;
+  openReportDialog: (categoryToOpen: ReportCategory) => void;
+  pickAttachments: (onImported?: (attachments: AgentAttachment[]) => void) => Promise<boolean>;
+  pickReportDialogAttachments: () => Promise<boolean>;
+  preferredVisionModel: VeniceModelDto | undefined;
+  proceedWithOversizeComposerInput: () => void;
+  removeAttachment: (id: string) => void;
+  removeReportDialogAttachment: (id: string) => void;
+  renderFundingNotice: ((context: TextFundingNoticeContext) => React.ReactNode) | undefined;
+  renderQueuedAttachmentFollowUp: (
+    queueKey: string,
+    item: QueuedAttachmentFollowUp,
+    options?: { demo?: boolean },
+  ) => JSX.Element;
+  renderSteerCard: (card: { id: string; text: string }) => JSX.Element;
+  reportDialogAttachments: ReportDialogAttachment[];
+  reportDialogCategory: ReportCategory;
+  reportDialogDescription: string;
+  reportDialogOpen: boolean;
+  restoreComposerDraft: (key: string | null) => void;
+  sandboxFirstItemRef: React.MutableRefObject<HTMLButtonElement | null>;
+  sandboxMenuOpen: boolean;
+  sandboxMenuRef: React.MutableRefObject<HTMLDivElement | null>;
+  sandboxTriggerRef: React.MutableRefObject<HTMLButtonElement | null>;
+  scrollTranscriptToLatest: () => void;
+  seedComposerNoteRef: (options?: { defer?: boolean }) => void;
+  selectedFollowUpCount: number;
+  selectedHermesSessionId: string | undefined;
+  selectedHermesSessionIsProvisional: boolean;
+  selectedQueuedAttachmentFollowUps: QueuedAttachmentFollowUp[];
+  selectedSteerCards: { id: string; text: string }[];
+  selectedTask: AgentTaskDto | undefined;
+  selectedUpNextDemoFollowUps: QueuedAttachmentFollowUp[];
+  sendReviewableIssueReport: (sessionId: string) => Promise<void>;
+  setAttachMenuOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  setCategory: React.Dispatch<React.SetStateAction<ReportCategory | null>>;
+  setComposerModelFlyout: React.Dispatch<React.SetStateAction<ModelPickerFlyout>>;
+  setComposerModelOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  setConfirmUnrestricted: React.Dispatch<React.SetStateAction<boolean>>;
+  setDraft: React.Dispatch<React.SetStateAction<string>>;
+  setDropActive: React.Dispatch<React.SetStateAction<boolean>>;
+  setFullModeDraft: React.Dispatch<React.SetStateAction<boolean>>;
+  setModelRootSearch: React.Dispatch<React.SetStateAction<string>>;
+  setModelSearch: React.Dispatch<React.SetStateAction<string>>;
+  setReportDialogCategory: React.Dispatch<React.SetStateAction<ReportCategory>>;
+  setReportDialogDescription: React.Dispatch<React.SetStateAction<string>>;
+  setReportDialogOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  setSandboxMenuOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  setSteerQueueOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  showImageModelWarning: boolean;
+  skillCommandLoading: boolean;
+  skills: HermesSkillInfo[] | null;
+  startDictation: () => Promise<void>;
+  steerCardsFade: ScrollFade;
+  steerCardsListRef: React.MutableRefObject<HTMLDivElement | null>;
+  steerQueueOpen: boolean;
+  stopHermesSession: (sessionId: string) => Promise<void>;
+  stoppingSessionIds: ReadonlySet<string>;
+  submit: (event?: FormEvent) => Promise<void>;
+  submitting: boolean;
+  switchOversizeComposerModel: () => void;
+  textActionsDisabledReason: string | undefined;
+  textFundingContext: TextFundingModelContext;
+  veniceApiKeyConfigured: boolean;
+  visibleComposerSizeWarning: ComposerInputSizeWarning | null;
+  visibleFollowUpQueueKey: string | undefined;
+  visibleIssueReportHasUnsentContext: boolean;
+  visibleIssueReportImportingFiles: boolean;
+  visibleIssueReportReview:
+    | { report: PendingIssueReport; sessionId: string; submitting: boolean }
+    | undefined;
+  workingSessionIds: Set<string>;
+};
